Fix buffer overflows, garbage data
Fix several issues with error checking and string handling.

1. The length of the atom, not the destination buffer, is used for
copying arguments.

1> A = list_to_atom(string:chars($x, 255)).
2> ossp_uuid:make(A,A).
Segmentation fault

2. text formatted UUID's return garbage (leak VM internal data)

The length of the buffer holding the UUID is too small. Although
uuid_export() returned an error (UUID_RC_MEM), the return value is
not checked. This results in a binary containing unzero'ed data being
returned to the caller.

3. ossp_uuid_nif_import can return garbage

In the case where an improperly sized UUID is passed to import/2,
uuid_import() is not used and garbage is returned to the caller.

4. Allocation is not checked

Check memory allocation succeeds to avoid overflow.

5. make_uuid_ns: allocate space for terminating NULL

Allow enif_get_string() to NULL terminate the strings.

All eunit tests pass now.